chore: mooore fixes
